ds_requirements.txt
scikit-learn==0.24.1
python-dotenv==0.15.0
pymongo==3.11.3
sentry-sdk==1.0.0
